requirements.txt
huggingface_hub
datasets
openai
numpy
pandas
python-dotenv
